requirements.txt
numpy~=1.23.5
pyqubo~=1.4.0
dwave-neal~=0.6.0
sympy~=1.11.1